Update README.md
Hi there! 👋
I have developed this to imporve my JS skills while learning new stuffs about it. It's pretty simple but includes some basic elements. Let's talk about them.

➡ Main Screen
You need to write your name there and there are some rules. Regex controls those boxes and not allowing numbers. Except that you cannot leave the boxes blank.
➡Generating Pin
After typing the name and surname, a pin is created with the initials of these strings and three numbers. Here is another control mechanism working. When you enter any other information other than the pin, it is not accepted. After the pin is entered correctly, the timeout function works and after 10 seconds, the page refreshes and disables the account.
➡Future updates
I will work on not deleting the account even if the page is refreshed using the local data of the browser.